(WIP)feat(gui-editor): use vim thorugh fyne terminal
By using the fyne terminal and executing (neo)vim, we could get the editor to be fit for coding purposes (shortcuts, syntax highlighting etc). As of now though, both editors behave out of the ordinary when executed in the fyne terminal. In addition to that there are issues with fsnotify and file saving, hence the WIP tag.