feat(code-entry): an entry wrapper for code editing
Similar to editor.go this is a wrapper around widget.Entry but it handles Ctrl/Cmd + S correctly and follows better implementation principles. The reason for making it a separate component for now is the planned addition of nvim and resulting refactoring of editor.go to become a container for the two.